+++ README.md
@@ -82,7 +82,7 @@ source = 0
 # La valeur par défaut est `0,1` (personne, vélo)
 classes = 0,1
 # La valeur par défaut est `1`, si vous voulez augmenter le temps entre chaque prise, augmentez la valeur
-interval = 0
+interval = 1
 # La valeur par défaut est `False`, si vous voulez activer l'affichage graphique, mettez `True`
 display_detection = False
 # La valeur par défaut est `False`, si vous voulez activer l'affichage des messages, mettez `True`
